microservices/python/knowledge-graph/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
neo4j==5.16.0
nats-py==2.6.0
python-dotenv==1.0.0
loguru==0.7.2
asyncio==3.4.3
numpy==1.26.3
tenacity==8.2.3
prometheus-client==0.17.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
pybreaker==1.0.1
-e ../common/nats_lib
